requirements.txt
# ── Streamlit backend (NGO dashboard) ─────────────────────────────────────────
streamlit>=1.35.0
openai>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0

# ── FastAPI / BioBERT service (see services/biobert/requirements.txt) ─────────
# Listed here for reference — install via services/biobert/requirements.txt
# fastapi==0.115.0
# uvicorn[standard]==0.30.6
# transformers==4.52.4
# torch==2.7.1

# ── Health check endpoint for Streamlit ───────────────────────────────────────
# Streamlit exposes /_stcore/health natively — no extra package needed
